edi: fix consumer mixin tests
