+++ edi/tests/__init__.py
@@ -3,3 +3,4 @@ from . import test_record
 from . import test_backend_base
 from . import test_backend_output
 from . import test_backend_input
+from . import test_consumer_mixin
